+++ bridges/bridge-template/Cargo.toml
@@ -38,4 +38,6 @@ postage  = "0.4"
 
 support-initialize = { path = "../../supports/support-initialize" }
 support-lifeline   = { path = "../../supports/support-lifeline" }
+support-terminal   = { path = "../../supports/support-terminal" }
+support-config     = { path = "../../supports/support-config" }
