+++ bridges/bridge-template/src/bridge/bus.rs
@@ -1,0 +1,3 @@
+use lifeline::prelude::*;
+
+lifeline_bus!(pub struct TemplateTaskBus);
