+++ bridges/bridge-template/src/bridge/mod.rs
@@ -1,0 +1,9 @@
+pub use self::bus::*;
+pub use self::config::*;
+pub use self::message::*;
+pub use self::task::*;
+
+mod bus;
+mod config;
+mod message;
+mod task;
