+++ bridges/bridge-template/src/bridge/task.rs
@@ -1,9 +1,6 @@
-use lifeline::dyn_bus::DynBus;
-
 use support_lifeline::task::TaskStack;
 
-use crate::bus::TemplateTaskBus;
-use crate::config::TemplateTaskConfig;
+use crate::bridge::TemplateTaskBus;
 use crate::service::some::SomeService;
 
 #[derive(Debug)]
@@ -11,17 +8,17 @@ pub struct TemplateTask {
     stack: TaskStack<TemplateTaskBus>,
 }
 
-impl BridgeTask<TemplateTaskBus> for TemplateTask {
-    fn stack(&mut self) -> &mut TaskStack<TemplateTaskBus> {
-        &mut self.stack
-    }
-}
-
 impl TemplateTask {
-    pub fn new(config: TemplateTaskConfig) -> color_eyre::Result<Self> {
+    pub fn new() -> color_eyre::Result<Self> {
         let bus = TemplateTaskBus::default();
         let mut stack = TaskStack::new(bus);
         stack.spawn_service::<SomeService>()?;
         Ok(Self { stack })
+    }
+}
+
+impl TemplateTask {
+    pub fn stack(&self) -> &TaskStack<TemplateTaskBus> {
+        &self.stack
     }
 }
