--- bridges/bridge-template/src/bus.rs
@@ -1,7 +1,0 @@
-use lifeline::prelude::*;
-use lifeline::{Receiver, Sender};
-
-use crate::message::ToTemplateLinkedMessage;
-use crate::task::TemplateTask;
-
-lifeline_bus!(pub struct TemplateTaskBus);
