+++ bridges/bridge-template/src/command/mod.rs
@@ -1,1 +1,2 @@
+pub mod handler;
 pub mod types;
