+++ bridges/bridge-template/src/command/types.rs
@@ -1,5 +1,14 @@
+use structopt::StructOpt;
+use support_terminal::types::BasicOptions;
+
+/// Bridge template options
 #[derive(Debug, StructOpt)]
 #[structopt(name = "bridge-template", about = "Bridge template")]
 pub enum Opts {
-    Start,
+    /// Start bridge template
+    Start {
+        /// basic options
+        #[structopt(flatten)]
+        basic_options: BasicOptions,
+    },
 }
