+++ bridges/bridge-template/src/main.rs
@@ -1,11 +1,16 @@
-mod bus;
+use structopt::StructOpt;
+
+use crate::command::types::Opts;
+
+mod bridge;
+mod cli;
 mod command;
-mod message;
 mod service;
-mod task;
 
 #[tokio::main]
 async fn main() -> color_eyre::Result<()> {
     support_initialize::init()?;
+    let opt = Opts::from_args();
+    cli::execute(opt).await?;
     Ok(())
 }
