+++ supports/support-config/src/types.rs
@@ -1,0 +1,8 @@
+use serde::{Deserialize, Serialize};
+
+/// Http client config
+#[derive(Clone, Debug, Default, Serialize, Deserialize)]
+pub struct HttpClientConfig {
+    /// The timeout of request
+    pub timeout: u64,
+}
