+++ supports/support-terminal/src/lib.rs
@@ -1,1 +1,3 @@
+pub mod helpers;
 pub mod output;
+pub mod types;
