+++ supports/support-terminal/src/types/mod.rs
@@ -1,0 +1,3 @@
+pub use self::command::*;
+
+mod command;
